requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
aiosqlite==0.19.0
supabase==2.3.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI and ML
httpx==0.26.0
google-generativeai==0.3.2
replicate==0.22.0

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80

# Environment and Configuration
python-dotenv==1.0.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0

# Hardware (for Raspberry Pi - optional)
gpiozero==2.0.1; platform_machine == "armv7l" or platform_machine == "aarch64"
picamera2==0.3.16; platform_machine == "armv7l" or platform_machine == "aarch64"

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Code Quality
flake8==7.0.0
black==24.1.1

# CORS
python-cors==1.0.0
